--- src/main/java/example/com/meneymergebe/domain/user/entity/User.java
@@ -1,29 +1,0 @@
-package example.com.meneymergebe.domain.user.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "tb_user")
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String nickname;
-
-    private String email;
-
-    private String profileUrl;
-
-    private UserRole role;
-
-}
