--- src/main/java/example/com/meneymergebe/domain/user/entity/UserRole.java
@@ -1,14 +1,0 @@
-package example.com.meneymergebe.domain.user.entity;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum UserRole {
-    ADMIN("ROLE_ADMIN", "관리자"),
-    USER("ROLE_USER", "사용자");
-
-    private final String authority;
-    private final String value;
-}
