--- src/main/java/example/com/meneymergebe/global/jwt/TokenStatus.java
@@ -1,5 +1,0 @@
-package example.com.meneymergebe.global.jwt;
-
-public enum TokenStatus {
-    INVALID, EXPIRED, VALID;
-}
